Review the pgcopydb commands.
  pgcopydb
    clone    Clone an entire database from source to target
    fork     Clone an entire database from source to target
    follow   Replay changes from the source database to the target database
    copy-db  Copy an entire database from source to target
