+++ src/bin/pgcopydb/cli_copy.c
@@ -19,11 +19,6 @@
 #include "string_utils.h"
 #include "summary.h"
 
-CopyDBOptions copyDBoptions = { 0 };
-
-static int cli_copy_db_getopts(int argc, char **argv);
-
-static void cli_copy_db(int argc, char **argv);
 static void cli_copy_schema(int argc, char **argv);
 static void cli_copy_data(int argc, char **argv);
 static void cli_copy_table_data(int argc, char **argv);
@@ -32,10 +27,7 @@ static void cli_copy_indexes(int argc, char **argv);
 static void cli_copy_constraints(int argc, char **argv);
 static void cli_copy_blobs(int argc, char **argv);
 
-static void cli_copy_prepare_specs(CopyDataSpec *copySpecs,
-								   CopyDataSection section);
-
-
+/* pgcopydb copy db is an alias for pgcopydb clone */
 CommandLine copy__db_command =
 	make_command(
 		"copy-db",
@@ -43,7 +35,6 @@ CommandLine copy__db_command =
 		" --source ... --target ... [ --table-jobs ... --index-jobs ... ] ",
 		"  --source              Postgres URI to the source database\n"
 		"  --target              Postgres URI to the target database\n"
-		"  --dir                 Work directory to use\n"
 		"  --table-jobs          Number of concurrent COPY jobs to run\n"
 		"  --index-jobs          Number of concurrent CREATE INDEX jobs to run\n"
 		"  --drop-if-exists      On the target database, clean-up from a previous run first\n"
@@ -57,9 +48,8 @@ CommandLine copy__db_command =
 		"  --not-consistent      Allow taking a new snapshot on the source database\n"
 		"  --snapshot            Use snapshot obtained with pg_export_snapshot\n",
 		cli_copy_db_getopts,
-		cli_copy_db);
+		cli_clone);
 
-/* have pgcopydb copy-db and pgcopydb copy db aliases to each-other */
 static CommandLine copy_db_command =
 	make_command(
 		"db",
@@ -80,7 +70,7 @@ static CommandLine copy_db_command =
 		"  --not-consistent      Allow taking a new snapshot on the source database\n"
 		"  --snapshot            Use snapshot obtained with pg_export_snapshot\n",
 		cli_copy_db_getopts,
-		cli_copy_db);
+		cli_clone);
 
 static CommandLine copy_schema_command =
 	make_command(
@@ -217,371 +207,6 @@ CommandLine copy_commands =
 	make_command_set("copy",
 					 "Implement the data section of the database copy",
 					 NULL, NULL, NULL, copy_subcommands);
-
-
-/*
- * cli_copy_db_getopts parses the CLI options for the `copy db` command.
- */
-static int
-cli_copy_db_getopts(int argc, char **argv)
-{
-	CopyDBOptions options = { 0 };
-	int c, option_index = 0, errors = 0;
-	int verboseCount = 0;
-
-	static struct option long_options[] = {
-		{ "source", required_argument, NULL, 'S' },
-		{ "target", required_argument, NULL, 'T' },
-		{ "dir", required_argument, NULL, 'D' },
-		{ "jobs", required_argument, NULL, 'J' },
-		{ "table-jobs", required_argument, NULL, 'J' },
-		{ "index-jobs", required_argument, NULL, 'I' },
-		{ "drop-if-exists", no_argument, NULL, 'c' }, /* pg_restore -c */
-		{ "no-owner", no_argument, NULL, 'O' },       /* pg_restore -O */
-		{ "no-comments", no_argument, NULL, 'X' },
-		{ "no-acl", no_argument, NULL, 'x' }, /* pg_restore -x */
-		{ "skip-blobs", no_argument, NULL, 'B' },
-		{ "skip-large-objects", no_argument, NULL, 'B' },
-		{ "filter", required_argument, NULL, 'F' },
-		{ "filters", required_argument, NULL, 'F' },
-		{ "restart", no_argument, NULL, 'r' },
-		{ "resume", no_argument, NULL, 'R' },
-		{ "not-consistent", no_argument, NULL, 'C' },
-		{ "snapshot", required_argument, NULL, 'N' },
-		{ "version", no_argument, NULL, 'V' },
-		{ "verbose", no_argument, NULL, 'v' },
-		{ "quiet", no_argument, NULL, 'q' },
-		{ "help", no_argument, NULL, 'h' },
-		{ NULL, 0, NULL, 0 }
-	};
-
-	optind = 0;
-
-	/* install default values */
-	options.tableJobs = 4;
-	options.indexJobs = 4;
-
-	/* read values from the environment */
-	if (!cli_copydb_getenv(&options))
-	{
-		log_fatal("Failed to read default values from the environment");
-		exit(EXIT_CODE_BAD_ARGS);
-	}
-
-	while ((c = getopt_long(argc, argv, "S:T:D:J:I:cOBrRCN:xXVvqh",
-							long_options, &option_index)) != -1)
-	{
-		switch (c)
-		{
-			case 'S':
-			{
-				if (!validate_connection_string(optarg))
-				{
-					log_fatal("Failed to parse --source connection string, "
-							  "see above for details.");
-					++errors;
-				}
-				strlcpy(options.source_pguri, optarg, MAXCONNINFO);
-				log_trace("--source %s", options.source_pguri);
-				break;
-			}
-
-			case 'T':
-			{
-				if (!validate_connection_string(optarg))
-				{
-					log_fatal("Failed to parse --target connection string, "
-							  "see above for details.");
-					++errors;
-				}
-				strlcpy(options.target_pguri, optarg, MAXCONNINFO);
-				log_trace("--target %s", options.target_pguri);
-				break;
-			}
-
-			case 'D':
-			{
-				strlcpy(options.dir, optarg, MAXPGPATH);
-				log_trace("--dir %s", options.dir);
-				break;
-			}
-
-			case 'J':
-			{
-				if (!stringToInt(optarg, &options.tableJobs) ||
-					options.tableJobs < 1 ||
-					options.tableJobs > 128)
-				{
-					log_fatal("Failed to parse --jobs count: \"%s\"", optarg);
-					++errors;
-				}
-				log_trace("--table-jobs %d", options.tableJobs);
-				break;
-			}
-
-			case 'I':
-			{
-				if (!stringToInt(optarg, &options.indexJobs) ||
-					options.indexJobs < 1 ||
-					options.indexJobs > 128)
-				{
-					log_fatal("Failed to parse --index-jobs count: \"%s\"", optarg);
-					++errors;
-				}
-				log_trace("--jobs %d", options.indexJobs);
-				break;
-			}
-
-			case 'c':
-			{
-				options.restoreOptions.dropIfExists = true;
-				log_trace("--drop-if-exists");
-				break;
-			}
-
-			case 'O':
-			{
-				options.restoreOptions.noOwner = true;
-				log_trace("--no-owner");
-				break;
-			}
-
-			case 'x':
-			{
-				options.restoreOptions.noACL = true;
-				log_trace("--no-ack");
-				break;
-			}
-
-			case 'X':
-			{
-				options.restoreOptions.noComments = true;
-				log_trace("--no-comments");
-				break;
-			}
-
-			case 'B':
-			{
-				options.skipLargeObjects = true;
-				log_trace("--skip-large-objects");
-				break;
-			}
-
-			case 'r':
-			{
-				options.restart = true;
-				log_trace("--restart");
-
-				if (options.resume)
-				{
-					log_fatal("Options --resume and --restart are not compatible");
-				}
-				break;
-			}
-
-			case 'R':
-			{
-				options.resume = true;
-				log_trace("--resume");
-
-				if (options.restart)
-				{
-					log_fatal("Options --resume and --restart are not compatible");
-				}
-				break;
-			}
-
-			case 'C':
-			{
-				options.notConsistent = true;
-				log_trace("--not-consistent");
-				break;
-			}
-
-			case 'N':
-			{
-				strlcpy(options.snapshot, optarg, sizeof(options.snapshot));
-				log_trace("--snapshot %s", options.snapshot);
-				break;
-			}
-
-			case 'F':
-			{
-				strlcpy(options.filterFileName, optarg, MAXPGPATH);
-				log_trace("--filters \"%s\"", options.filterFileName);
-
-				if (!file_exists(options.filterFileName))
-				{
-					log_error("Filters file \"%s\" does not exists",
-							  options.filterFileName);
-					++errors;
-				}
-				break;
-			}
-
-			case 'V':
-			{
-				/* keeper_cli_print_version prints version and exits. */
-				cli_print_version(argc, argv);
-				break;
-			}
-
-			case 'v':
-			{
-				++verboseCount;
-				switch (verboseCount)
-				{
-					case 1:
-					{
-						log_set_level(LOG_INFO);
-						break;
-					}
-
-					case 2:
-					{
-						log_set_level(LOG_DEBUG);
-						break;
-					}
-
-					default:
-					{
-						log_set_level(LOG_TRACE);
-						break;
-					}
-				}
-				break;
-			}
-
-			case 'q':
-			{
-				log_set_level(LOG_ERROR);
-				break;
-			}
-
-			case 'h':
-			{
-				commandline_help(stderr);
-				exit(EXIT_CODE_QUIT);
-				break;
-			}
-		}
-	}
-
-	if (IS_EMPTY_STRING_BUFFER(options.source_pguri) ||
-		IS_EMPTY_STRING_BUFFER(options.target_pguri))
-	{
-		log_fatal("Options --source and --target are mandatory");
-		exit(EXIT_CODE_BAD_ARGS);
-	}
-
-	if (!cli_copydb_is_consistent(&options))
-	{
-		log_fatal("Option --resume requires option --not-consistent");
-		exit(EXIT_CODE_BAD_ARGS);
-	}
-
-	if (errors > 0)
-	{
-		commandline_help(stderr);
-		exit(EXIT_CODE_BAD_ARGS);
-	}
-
-	/* publish our option parsing in the global variable */
-	copyDBoptions = options;
-
-	return optind;
-}
-
-
-/*
- * cli_copy_db implements the command: pgcopydb copy db
- */
-static void
-cli_copy_db(int argc, char **argv)
-{
-	CopyDataSpec copySpecs = { 0 };
-
-	(void) cli_copy_prepare_specs(&copySpecs, DATA_SECTION_ALL);
-
-	Summary summary = { 0 };
-	TopLevelTimings *timings = &(summary.timings);
-
-	(void) summary_set_current_time(timings, TIMING_STEP_START);
-
-	log_info("STEP 1: dump the source database schema (pre/post data)");
-
-	(void) summary_set_current_time(timings, TIMING_STEP_BEFORE_SCHEMA_DUMP);
-
-	/*
-	 * First, we need to open a snapshot that we're going to re-use in all our
-	 * connections to the source database. When the --snapshot option has been
-	 * used, instead of exporting a new snapshot, we can just re-use it.
-	 */
-	if (!copydb_prepare_snapshot(&copySpecs))
-	{
-		/* errors have already been logged */
-		exit(EXIT_CODE_INTERNAL_ERROR);
-	}
-
-	if (!copydb_dump_source_schema(&copySpecs,
-								   copySpecs.sourceSnapshot.snapshot,
-								   PG_DUMP_SECTION_SCHEMA))
-	{
-		/* errors have already been logged */
-		(void) copydb_close_snapshot(&copySpecs);
-		exit(EXIT_CODE_INTERNAL_ERROR);
-	}
-
-	log_info("STEP 2: restore the pre-data section to the target database");
-
-	/* fetch schema information from source catalogs, including filtering */
-	if (!copydb_fetch_schema_and_prepare_specs(&copySpecs))
-	{
-		/* errors have already been logged */
-		(void) copydb_close_snapshot(&copySpecs);
-		exit(EXIT_CODE_TARGET);
-	}
-
-	(void) summary_set_current_time(timings, TIMING_STEP_BEFORE_PREPARE_SCHEMA);
-
-	if (!copydb_target_prepare_schema(&copySpecs))
-	{
-		/* errors have already been logged */
-		(void) copydb_close_snapshot(&copySpecs);
-		exit(EXIT_CODE_TARGET);
-	}
-
-	(void) summary_set_current_time(timings, TIMING_STEP_AFTER_PREPARE_SCHEMA);
-
-	log_info("STEP 3: copy data from source to target in sub-processes");
-	log_info("STEP 4: create indexes and constraints in parallel");
-	log_info("STEP 5: vacuum analyze each table");
-
-	if (!copydb_copy_all_table_data(&copySpecs))
-	{
-		/* errors have already been logged */
-		(void) copydb_close_snapshot(&copySpecs);
-		exit(EXIT_CODE_INTERNAL_ERROR);
-	}
-
-	/* now close the snapshot we kept for the whole operation */
-	(void) copydb_close_snapshot(&copySpecs);
-
-	log_info("STEP 7: restore the post-data section to the target database");
-
-	(void) summary_set_current_time(timings, TIMING_STEP_BEFORE_FINALIZE_SCHEMA);
-
-	if (!copydb_target_finalize_schema(&copySpecs))
-	{
-		/* errors have already been logged */
-		exit(EXIT_CODE_TARGET);
-	}
-
-	(void) summary_set_current_time(timings, TIMING_STEP_AFTER_FINALIZE_SCHEMA);
-	(void) summary_set_current_time(timings, TIMING_STEP_END);
-
-	(void) print_summary(&summary, &copySpecs);
-}
 
 
 /*
@@ -990,80 +615,4 @@ cli_copy_blobs(int argc, char **argv)
 
 	(void) summary_set_current_time(timings, TIMING_STEP_END);
 	(void) print_summary(&summary, &copySpecs);
-}
-
-
-/*
- * cli_copy_prepare_specs initializes our internal data structure that are used
- * to drive the operations.
- */
-static void
-cli_copy_prepare_specs(CopyDataSpec *copySpecs, CopyDataSection section)
-{
-	PostgresPaths *pgPaths = &(copySpecs->pgPaths);
-
-	char scrubbedSourceURI[MAXCONNINFO] = { 0 };
-	char scrubbedTargetURI[MAXCONNINFO] = { 0 };
-
-	(void) parse_and_scrub_connection_string(copyDBoptions.source_pguri,
-											 scrubbedSourceURI);
-
-	(void) parse_and_scrub_connection_string(copyDBoptions.target_pguri,
-											 scrubbedTargetURI);
-
-	log_info("[SOURCE] Copying database from \"%s\"", scrubbedSourceURI);
-	log_info("[TARGET] Copying database into \"%s\"", scrubbedTargetURI);
-
-	(void) find_pg_commands(pgPaths);
-
-	log_debug("Using pg_dump for Postgres \"%s\" at \"%s\"",
-			  copySpecs->pgPaths.pg_version,
-			  copySpecs->pgPaths.pg_dump);
-
-	log_debug("Using pg_restore for Postgres \"%s\" at \"%s\"",
-			  copySpecs->pgPaths.pg_version,
-			  copySpecs->pgPaths.pg_restore);
-
-	char *dir =
-		IS_EMPTY_STRING_BUFFER(copyDBoptions.dir)
-		? NULL
-		: copyDBoptions.dir;
-
-	if (!copydb_init_workdir(copySpecs,
-							 dir,
-							 copyDBoptions.restart,
-							 copyDBoptions.resume))
-	{
-		/* errors have already been logged */
-		exit(EXIT_CODE_INTERNAL_ERROR);
-	}
-
-	if (!copydb_init_specs(copySpecs,
-						   copyDBoptions.source_pguri,
-						   copyDBoptions.target_pguri,
-						   copyDBoptions.tableJobs,
-						   copyDBoptions.indexJobs,
-						   section,
-						   copyDBoptions.snapshot,
-						   copyDBoptions.restoreOptions,
-						   copyDBoptions.skipLargeObjects,
-						   copyDBoptions.restart,
-						   copyDBoptions.resume,
-						   !copyDBoptions.notConsistent))
-	{
-		/* errors have already been logged */
-		exit(EXIT_CODE_INTERNAL_ERROR);
-	}
-
-	if (!IS_EMPTY_STRING_BUFFER(copyDBoptions.filterFileName))
-	{
-		SourceFilters *filters = &(copySpecs->filters);
-
-		if (!parse_filters(copyDBoptions.filterFileName, filters))
-		{
-			log_error("Failed to parse filters in file \"%s\"",
-					  copyDBoptions.filterFileName);
-			exit(EXIT_CODE_BAD_ARGS);
-		}
-	}
 }
