+++ src/bin/pgcopydb/copydb.c
@@ -9,6 +9,7 @@
 #include <sys/wait.h>
 #include <unistd.h>
 
+#include "cli_common.h"
 #include "copydb.h"
 #include "env_utils.h"
 #include "lock_utils.h"
