+++ tests/filtering/copydb.sh
@@ -21,8 +21,8 @@ pgcopydb list tables --source ${PGCOPYDB_TARGET_PGURI}
 psql -d ${PGCOPYDB_SOURCE_PGURI} -1 -f /usr/src/pagila/pagila-schema.sql
 psql -d ${PGCOPYDB_SOURCE_PGURI} -1 -f /usr/src/pagila/pagila-data.sql
 
-# pgcopydb copy db uses the environment variables
-pgcopydb copy-db --filters /usr/src/pgcopydb/include.ini
+# pgcopydb clone uses the environment variables
+pgcopydb clone --filters /usr/src/pgcopydb/include.ini
 
 # now compare the output of running the SQL command with what's expected
 # as we're not root when running tests, can't write in /usr/src
